Add pydantic models for questions and other objects
But for now, maintain backwards compatibility with the old structured responses way of doing things.
Next step will be to change the server itself to return structured responses.

As part of this commit, I snuck in a change to pull the FATEBOOK_API_KEY in the test from the .env file.
Even though it's technically probably safe to just leave the test account key in there, I want to practice caution given this is a public repo.
(Any malicious users should know that the old ones from the Git history are all invalidated, so don't even bother.)