Migrate list_questions to structured responses
Had to do a bit more work here than expected because the list[Question] type wasn't working, so instead we made a new type, QuestionsList. But then to get MCP inspector to work, we needed to use the 'result' key for the list, which seems like some hidden assumption that I'd like to get to the bottom of later.

Anyway, now it works and we're ready to migrate others over.